+++ Caller-ID/src/main/res/layout/item_third_activity.xml
@@ -7,6 +7,7 @@
         android:id="@+id/item"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:paddingBottom="20dp"
         android:orientation="horizontal">
         <ImageView
             android:id="@+id/phone"
